Prepare/PreparePhase/requirements.txt
mysql-connector-python~=8.0.13
lxml~=5.2.2